+++ tests/test_genetics.py
@@ -1,0 +1,14 @@
+# -*- encoding: utf-8 -*-
+"""Tests for rizza.genetics."""
+import pytest
+from rizza.genetics import *
+
+
+@pytest.mark.skip(reason="Not currently implemented.")
+def test_positive_create_population():
+    pass
+
+
+@pytest.mark.skip(reason="Not currently implemented.")
+def test_positive_create_organism():
+    pass